Update README.md
# yPoemas

Hi... my name is Nando Lopes

I'm a young old-man (69 yo) from Brazil.

I have developed a software that ramdomly gathers
words and phrases and the result of such thing : a poetry is born.

It is coded in Python/StreamLit.
Since I am a " very beginner" in these languages,
I know that my code may look like a "Frankenstein".

Actually it runs in 6 different idioms (pt, es, it, fr, en, de)
for the generated text (by Google-Translator).

I'm looking for someone that could help me to revise the project.

I wonder if there is a good Godsended angel interested in such a job...

Like it? Contact me for more details.